refactor services with index
