+++ electron/services/devTools.ts
@@ -1,7 +1,7 @@
 import installExtension, {
   REACT_DEVELOPER_TOOLS,
 } from 'electron-devtools-installer';
-import { ILogger } from './services/logger';
+import { ILogger } from './logger';
 
 export function setUpDevtools(logger: ILogger) {
   return (): void => {
