+++ electron/services/index.ts
@@ -1,0 +1,3 @@
+export * from './devTools';
+export * from './logger';
+export * from './updater';
